Gigantic commit! By far the most important change is this: Ovum now uses status codes (of type int) to indicate error conditions during method evaluation and when invoking native functions, instead of throwing OvumException (which has been removed).
This change was made because native functions are declared with extern "C", to prevent name mangling, and throwing exceptions from a C++ function invoked through a C function causes undefined behaviour. It has worked on Windows mainly because Microsoft have gone to some length to make their compiler support it, but this is not in any way guaranteed to continue working with other compilers. Hence, this particular dependency on undefined behaviour has been removed. Note that some functions return nullptr instead of a status code, in which case nullptr indicates an out-of-memory condition.

An alternative solution would have been to use Microsoft's Structured Exception Handling (http://msdn.microsoft.com/en-us/library/windows/desktop/ms680657.aspx), which has been designed to work with both C and C++. But then Ovum would basically be doomed to remain Windows-only software forever, unless significant parts of the VM were rewritten (with macros or otherwise) for other systems, which would basically involve maintaining two code bases.

Since all NativeFunctions are now required to return status codes, every single native function in aves has been changed. The macros BEGIN_NATIVE_FUNCTION, END_NATIVE_FUNCTION, CHECKED and CHECKED_MEM have been added to aid in the writing of native functions.

Fair warning: it is entirely possible that I may have forgotten to check the return code in some cases. As always, don't rely on Ovum behaving sensibly; it's very much alpha software!

Some secondary changes:
[all]
* Enabled "Treat Warnings as Errors"

[VM]
* The implementations of MemberTable<T> and ModuleReader have been moved to their own files, to reduce the clutter of the still somewhat messy module headers.
* Checked integer arithmetic functions return a status code to indicate overflow/divide by zero conditions, and take a parameter to put the return value into. They no longer take a ThreadHandle.

[aves]
* Native function string::Replace actually works correctly (sometimes appended too many characters to output buffer)
* Reorganised string::Format (both versions, but especially the one that takes a Hash)

+++ VM/ov_thread.opcodes.h
@@ -420,14 +420,14 @@ namespace instr
 
 		inline uint32_t GetOriginalOffset(const int32_t index) const
 		{
-			if (index >= instructions.size())
+			if (index >= (int32_t)instructions.size())
 				return instructions.end()->originalOffset +
 					instructions.end()->originalSize;
 			return instructions[index].originalOffset;
 		}
 		inline uint32_t GetOriginalSize(const int32_t index) const
 		{
-			if (index >= instructions.size())
+			if (index >= (int32_t)instructions.size())
 				return 0;
 			return instructions[index].originalSize;
 		}
